+++ src/components/BrowseResultsList/utils.js
@@ -1,4 +1,7 @@
+import omit from 'lodash/omit';
+
 import {
+  browseCallNumberOptions,
   browseModeOptions,
   FACETS,
   queryIndexes,
@@ -14,45 +17,89 @@ export const isRowPreventsClick = (row, browseOption) => {
   );
 };
 
-export const getSearchParams = (row, qindex) => {
+const facetsToString = (filters, facetNameInBrowse, facetNameInSearch) => {
+  return filters[facetNameInBrowse]?.map(value => `${facetNameInSearch}.${value}`).join(',');
+};
+
+const getExtraFilters = (row, qindex, allFilters) => {
+  const filtersOnly = omit(allFilters, 'qindex', 'query');
+  const extraFacets = [];
+
+  let sharedFacetName;
+  let heldByFacetName;
+
+  if (qindex === browseModeOptions.SUBJECTS) {
+    sharedFacetName = FACETS.SUBJECTS_SHARED;
+    heldByFacetName = FACETS.SUBJECTS_HELD_BY;
+
+    if (row.authorityId) {
+      extraFacets.push(`${FACETS.AUTHORITY_ID}.${row.authorityId}`);
+    }
+  } else if (qindex === browseModeOptions.CONTRIBUTORS) {
+    sharedFacetName = FACETS.CONTRIBUTORS_SHARED;
+    heldByFacetName = FACETS.CONTRIBUTORS_HELD_BY;
+
+    extraFacets.push(`${FACETS.SEARCH_CONTRIBUTORS}.${row.contributorNameTypeId}`);
+  } else if (Object.values(browseCallNumberOptions).includes(qindex)) {
+    sharedFacetName = FACETS.SHARED;
+    heldByFacetName = FACETS.CALL_NUMBERS_HELD_BY;
+  }
+
+  const sharedExtraFacets = facetsToString(filtersOnly, sharedFacetName, FACETS.SHARED);
+  const heldByExtraFacets = facetsToString(filtersOnly, heldByFacetName, FACETS.HELD_BY);
+  const extraFacetsString = [...extraFacets, sharedExtraFacets, heldByExtraFacets].filter(Boolean).join(',');
+
+  return extraFacetsString ? { filters: extraFacetsString } : {};
+};
+
+export const getSearchParams = (row, qindex, allFilters) => {
+  const filters = getExtraFilters(row, qindex, allFilters);
+
   const optionsMap = {
     [browseModeOptions.CALL_NUMBERS]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.DEWEY]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.LIBRARY_OF_CONGRESS]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.LOCAL]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.NATIONAL_LIBRARY_OF_MEDICINE]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.OTHER]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.SUPERINTENDENT]: {
       qindex: queryIndexes.CALL_NUMBER,
       query: row.shelfKey,
+      ...filters,
     },
     [browseModeOptions.CONTRIBUTORS]: {
       qindex: queryIndexes.CONTRIBUTOR,
       query: row.name,
-      filters: `${FACETS.SEARCH_CONTRIBUTORS}.${row.contributorNameTypeId}`,
+      ...filters,
     },
     [browseModeOptions.SUBJECTS]: {
       qindex: queryIndexes.SUBJECT,
       query: row.value,
-      ...(row.authorityId && { filters: `${FACETS.AUTHORITY_ID}.${row.authorityId}` }),
+      ...filters,
     },
   };