+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Reset CheckboxFacet state.more when user resets search form and fewer facet options are loaded. Fixes UIIN-2531.
 * Edit instance success toast no longer shows the instance HRID. Fixes UIIN-2588.
 * Show facet options, if they exist, after clicking the +More button. Refs UIIN-2533.
+* If Shared & Held by facets were selected in the Browse search, then retain them in the Search lookup after clicking the record. Refs UIIN-2608.
 * Instance. Series heading has vanished in detailed view. Fixes UIIN-2601.
 
 ## [10.0.0](https://github.com/folio-org/ui-inventory/tree/v10.0.0) (2023-10-13)